Module fe_parser::builders

Parser factory functions (builders).

Much of the code in this module was inspired by the nom library. One or two particular parts of this module were copied more or less directly from nom. Therefore, the nom license is included below and will apply where explicitly noted. Otherwise, though the API of this module was directly taken from nom, the actual implementations of most parser builders are original, though they may coincide with the implementations of nom parser builders.

<<<< BEGIN NOM LICENSE >>>>

Copyright (c) 2014-2019 Geoffroy Couprie

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to
deal in the Software without restriction, including without limitation the
rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
sell copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
IN THE SOFTWARE.

<<<< END NOM LICENSE >>>>

The above license applies to any component enclosed in a pair of comments as follows:

// <<< BEGIN NOM LICENSE COVERAGE >>>

...

// <<< END NOM LICENSE COVERAGE >>>

Traits

Alt

Functions

alt

delimited
Modify a parser to parse and discard left and right delimiters before and after the parsed content. The parsed delimiters are used to determine the span of the parsing result.

many0

many1

map

op_expr_builder

opt

pair

preceded

separated
Modify a parser to apply itself repeatedly while consuming separators (including an optional trailing separator).

terminated

verify