Function fe_parser::builders::delimited

pub fn delimited<'a, L, P, R, O1, O2, O3>(
    left: L, 
    parser: P, 
    right: R
) -> impl Fn(Cursor<'a>) -> ParseResult<'_, Spanned<O2>> where
    L: Fn(Cursor<'a>) -> ParseResult<'_, O1>,
    P: Fn(Cursor<'a>) -> ParseResult<'_, O2>,
    R: Fn(Cursor<'a>) -> ParseResult<'_, O3>,
    O1: Into<Span>,
    O3: Into<Span>, 

Modify a parser to parse and discard left and right delimiters before and after the parsed content. The parsed delimiters are used to determine the span of the parsing result.