fe_parser::node

Struct Span

pub struct Span {
    pub file_id: SourceFileId,
    pub start: usize,
    pub end: usize,
}

An exclusive span of byte offsets in a source file.

Fields
file_id: SourceFileIdstart: usize

A byte offset specifying the inclusive start of a span.

end: usize

A byte offset specifying the exclusive end of a span.

Implementations

impl Span

pub fn new(file_id: SourceFileId, start: usize, end: usize) -> Span

pub fn zero(file_id: SourceFileId) -> Span

pub fn dummy() -> Span

pub fn is_dummy(&self) -> bool

pub fn from_pair<S, E>(start_elem: S, end_elem: E) -> Span
where S: Into<Span>, E: Into<Span>,