Module fe_parser::string_utils

Structs

Position
A position in a source file specified by a 1-indexed line number and a 0-indexed byte offset into the line specified by that number.

StringPositions
Efficiently find the text positions (line, column tuples) of a monotonically increasing sequence of byte offsets in a string. Non-monotonic sequences are also supported but are less efficient.

Functions

lines_with_endings
Iterate over the lines in buf and include line endings in the results. Also, provide byte offsets of line beginnings and endings.

lstrip_slice
Strip the characters in the string strip from the left side of the string slice input.

rstrip_slice
Strip the characters in the string strip from the right side of the string slice input.