Struct fe_parser::Parser

pub struct Parser<'a> {
    pub diagnostics: Vec<Diagnostic>,
    // some fields omitted
}

Parser maintains the parsing state, such as the token stream, indent stack, paren stack, diagnostics, etc. Syntax parsing logic is in the crate::grammar module.

See [BTParser] if you need backtrackable parser.

Fields
diagnostics: Vec<Diagnostic>

The diagnostics (errors and warnings) emitted during parsing.

Implementations

impl<'a> Parser<'a>

pub fn new(content: &'a str, file_id: SourceFileId) -> Self

Create a new parser for a source code string and associated file id.

pub fn as_bt_parser<'b>(&'b mut self) -> BTParser<'a, 'b>

Return as wrapped back tracking parser

pub fn next(&mut self) -> ParseResult<Token<'a>>

Return the next token, or an error if we’ve reached the end of the file.

pub fn peek_or_err(&mut self) -> ParseResult<TokenKind>

Take a peek at the next token kind without consuming it, or return an error if we’ve reached the end of the file.

pub fn peek(&mut self) -> Option<TokenKind>

Take a peek at the next token kind. Returns None if we’ve reached the end of the file.

pub fn peeked_text(&mut self) -> &'a str

Peek at the text of the next token, without consuming it. This is useful for words that are keywords in some contexts, but not others. E.g. from can be used as a variable or field name, but it’s the leading keyword of a from x import y statement.

Panics

This function must only be used in cases where the kind of the next token has already been Parser::peeked. If there is no next token, it will panic.

pub fn split_next(&mut self) -> ParseResult<Token<'a>>

Split the next token into two tokens, returning the first. Only supports splitting the >> token into two > tokens, specifically for parsing the closing angle bracket of a generic type argument list (Map<x, Map<y, z>>).

Panics

Panics if the next token isn’t >>

pub fn done(&mut self) -> bool

Returns true if the parser has reached the end of the file.

pub fn last_indent(&self) -> &'a str

The leading whitespace string of the last-seen indented line. This does not include lines inside of parentheses.

pub fn assert(&mut self, tk: TokenKind) -> Token<'a>

Assert that the next token kind it matches the expected token kind, and return it. This should be used in cases where the next token kind is expected to have been checked already.

Panics

Panics if the next token kind isn’t tk.

pub fn expect<S: Into<String>>(
&mut self,
expected: TokenKind,
message: S
) -> ParseResult<Token<'a>>

If the next token matches the expected kind, return it. Otherwise emit an error diagnostic with the given message and return an error.

pub fn expect_with_notes<Str, NotesFn>(
&mut self,
expected: TokenKind,
message: Str,
notes_fn: NotesFn
) -> ParseResult<Token<'a>> where
Str: Into<String>,
NotesFn: FnOnce() -> Vec<String>,

Like Parser::expect, but with additional notes to be appended to the bottom of the diagnostic message. The notes are provided by a function that returns a Vec<String>, to avoid allocations in the case where the token is as expected.

pub fn unexpected_token_error<S: Into<String>>(
&mut self,
span: Span,
message: S,
notes: Vec<String>
)

Emit an “unexpected token” error diagnostic with the given message.

pub fn enter_block(
&mut self,
context_span: Span,
context_name: &str
) -> ParseResult<()>

Enter an indented block, which is expected to be non-empty. This checks for and consumes the colon that precedes the block. If no colon is found, it emits an error and keeps parsing. Any number of newlines are allowed before the first non-empty indented line. Returns an error if the block has no non-empty indented line.

Panics

Panics if called while the parser is inside a set of parentheses.

pub fn expect_newline(&mut self, context_name: &str) -> ParseResult<()>

Expect and consume one or more newlines, without returning them. Returns an error if the next token isn’t a newline or the end of the file, or if the indent of the next non-empty line doesn’t match the current indentation.

Panics

Panics if called while the parser is inside a set of parentheses.

pub fn error<S: Into<String>>(&mut self, span: Span, message: S)

Emit an error diagnostic, but don’t stop parsing

pub fn fancy_error<S: Into<String>>(
&mut self,
message: S,
labels: Vec<Label>,
notes: Vec<String>
)

Emit a “fancy” error diagnostic with any number of labels and notes, but don’t stop parsing.